requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-openai
python-dotenv
bs4
pypdf
unstructured
fake_useragent
youtube-transcript-api==0.6.2